refactor(flow): add action step type; move flow types to own file
