+++ packages/core/src/types/index.ts
@@ -1,2 +1,3 @@
 export * from "./event-bus.types.js";
 export * from "./engine.types.js";
+export * from "./flow.types.js";
